+++ server/src/main/java/net/anyflow/lannister/Settings.java
@@ -1,12 +1,12 @@
 /*
  * Copyright 2016 The Lannister Project
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -34,7 +34,7 @@ public class Settings extends java.util.Properties {
 
 	protected static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(Settings.class);
 
-	private final static String CONFIG_NAME = "lannister.cfg";
+	private final static String CONFIG_NAME = "lannister.properties";
 
 	public final static Settings INSTANCE = new Settings();
 
@@ -161,6 +161,8 @@ public class Settings extends java.util.Properties {
 	}
 
 	public String nettyTransportMode() {
+		if (!getBoolean("netty.nativeTransportMode", true)) { return Literals.NETTY_NIO; }
+
 		return System.getProperty("os.name").startsWith("Linux") ? Literals.NETTY_EPOLL : Literals.NETTY_NIO;
 	}